Multimodal/Qwen3-VL-30B-A3B-Instruct_API_Vision_Chatbot_Example/requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
Pillow>=10.3.0
requests>=2.31.0
langchain>=0.1.0
